requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.7
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Markdown==3.5.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==1.26.18
